+++ pytzer/__init__.py
@@ -14,8 +14,8 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """Pitzer model for chemical activities in aqueous solutions."""
-from . import (cflibs, coefficients, constants, debyehueckel, dissociation, io,
-    jfuncs, matrix, meta, model, properties, tables, teos10)
+from . import (cflibs, coefficients, constants, debyehueckel, dissociation,
+    equilibrate, io, jfuncs, matrix, meta, model, properties, tables, teos10)
 
 __all__ = [
     'cflibs',
@@ -23,6 +23,7 @@ __all__ = [
     'constants',
     'debyehueckel',
     'dissociation',
+    'equilibrate',
     'io',
     'jfuncs',
     'matrix',